scripts/utils/ocr/requirements.txt
mistralai>=1.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
